pyproject.toml

[project]
name = "phoneme_contrast"
version = "0.1.0"
description = "Phoneme classification using contrastive learning"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "hydra-core>=1.3.0",
    "wandb>=0.15.0",
    "tensorboard>=2.13.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "soundfile>=0.13.1",
    "omegaconf>=2.3.0",
    "polars>=1.31.0",
    "ruff>=0.11.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "ruff>=0.0.270",
    "mypy>=1.3.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
]
